refactor: split the entities to multiple files.
