+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/FitnessGoal.kt
@@ -1,0 +1,7 @@
+package com.cairosquad.evolvefit.entity
+
+enum class FitnessGoal {
+    LOSE_WEIGHT,
+    GAIN_WEIGHT,
+    STAY_IN_SHAPE
+}
