+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/FocusArea.kt
@@ -1,0 +1,10 @@
+package com.cairosquad.evolvefit.entity
+
+enum class FocusArea {
+    Quadriceps,
+    Abs,
+    Calves,
+    LowerBack,
+    Core,
+    Shoulders
+}
