+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/Gender.kt
@@ -1,0 +1,6 @@
+package com.cairosquad.evolvefit.entity
+
+enum class Gender {
+    MALE,
+    FEMALE
+}
