+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/MeasurementType.kt
@@ -1,0 +1,6 @@
+package com.cairosquad.evolvefit.entity
+
+enum class MeasurementType {
+    DURATION,
+    REPS
+}
