+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/MeasurementUnit.kt
@@ -1,0 +1,6 @@
+package com.cairosquad.evolvefit.entity
+
+enum class MeasurementUnit {
+    METRIC,
+    IMPERIAL
+}
