+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/entity/WorkoutDay.kt
@@ -1,0 +1,11 @@
+package com.cairosquad.evolvefit.entity
+
+enum class WorkoutDay {
+    MONDAY,
+    TUESDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY,
+    SUNDAY
+}
